// src/astro/models/Chart.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { ChartDate } from './ChartDate';
import type { Position } from './Position';

export class Chart {
    'chartDate': ChartDate;
    'sun': Position;
    'moon': Position;
    'northNode': Position;
    'mercury': Position;
    'venus': Position;
    'mars': Position;
    'jupiter': Position;
    'saturn': Position;
    'chiron': Position;
    'uranus': Position;
    'neptune': Position;
    'pluto': Position;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "chartDate",
            "baseName": "chart_date",
            "type": "ChartDate",
            "format": ""
        },
        {
            "name": "sun",
            "baseName": "sun",
            "type": "Position",
            "format": ""
        },
        {
            "name": "moon",
            "baseName": "moon",
            "type": "Position",
            "format": ""
        },
        {
            "name": "northNode",
            "baseName": "north_node",
            "type": "Position",
            "format": ""
        },
        {
            "name": "mercury",
            "baseName": "mercury",
            "type": "Position",
            "format": ""
        },
        {
            "name": "venus",
            "baseName": "venus",
            "type": "Position",
            "format": ""
        },
        {
            "name": "mars",
            "baseName": "mars",
            "type": "Position",
            "format": ""
        },
        {
            "name": "jupiter",
            "baseName": "jupiter",
            "type": "Position",
            "format": ""
        },
        {
            "name": "saturn",
            "baseName": "saturn",
            "type": "Position",
            "format": ""
        },
        {
            "name": "chiron",
            "baseName": "chiron",
            "type": "Position",
            "format": ""
        },
        {
            "name": "uranus",
            "baseName": "uranus",
            "type": "Position",
            "format": ""
        },
        {
            "name": "neptune",
            "baseName": "neptune",
            "type": "Position",
            "format": ""
        },
        {
            "name": "pluto",
            "baseName": "pluto",
            "type": "Position",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Chart.attributeTypeMap;
    }

    public constructor() {
    }
}